refactor: separate OMG.State-dependent tests to other file
